Moved ExecutorConnection to correct package
